+++ intercom_flutter/test/intercom_flutter_test.dart
@@ -283,4 +283,48 @@ void main() {
     await Intercom.instance.displayHome();
     expectMethodCall('displayHome');
   });
+
+  test('setUserJwt', () {
+    Intercom.instance.setUserJwt('test');
+    expectMethodCall('setUserJwt', arguments: {
+      'jwt': 'test',
+    });
+  });
+
+  test('setAuthTokens', () {
+    Intercom.instance.setAuthTokens({'security_token': 'test'});
+    expectMethodCall('setAuthTokens', arguments: {
+      'tokens': {'security_token': 'test'},
+    });
+  });
+
+  group('setThemeMode', () {
+    test('dark', () {
+      Intercom.instance.setThemeMode(IntercomTheme.dark);
+      expectMethodCall('setThemeMode', arguments: {
+        'theme': 'dark',
+      });
+    });
+
+    test('light', () {
+      Intercom.instance.setThemeMode(IntercomTheme.light);
+      expectMethodCall('setThemeMode', arguments: {
+        'theme': 'light',
+      });
+    });
+
+    test('system', () {
+      Intercom.instance.setThemeMode(IntercomTheme.system);
+      expectMethodCall('setThemeMode', arguments: {
+        'theme': 'system',
+      });
+    });
+
+    test('none', () {
+      Intercom.instance.setThemeMode(IntercomTheme.none);
+      expectMethodCall('setThemeMode', arguments: {
+        'theme': 'none',
+      });
+    });
+  });
 }