+++ intercom_flutter_platform_interface/lib/method_channel_intercom_flutter.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/services.dart';
+import 'package:intercom_flutter_platform_interface/enumeral.dart';
 import 'package:intercom_flutter_platform_interface/intercom_flutter_platform_interface.dart';
 import 'package:intercom_flutter_platform_interface/intercom_status_callback.dart';
 
@@ -174,6 +175,8 @@ class MethodChannelIntercomFlutter extends IntercomFlutterPlatform {
     await _channel.invokeMethod('sendTokenToIntercom', {'token': token});
   }
 
+  @Deprecated(
+      "Calling this API is no longer required. Intercom will directly open the chat screen when a push notification is clicked.")
   @override
   Future<void> handlePushMessage() async {
     await _channel.invokeMethod('handlePushMessage');
@@ -251,6 +254,21 @@ class MethodChannelIntercomFlutter extends IntercomFlutterPlatform {
     var attributes = Map<String, dynamic>.from(
         await _channel.invokeMethod<Map>('fetchLoggedInUserAttributes') ?? {});
     return attributes;
+  }
+
+  @override
+  Future<void> setUserJwt(String jwt) async {
+    await _channel.invokeMethod('setUserJwt', {'jwt': jwt});
+  }
+
+  @override
+  Future<void> setAuthTokens(Map<String, String> tokens) async {
+    await _channel.invokeMethod('setAuthTokens', {'tokens': tokens});
+  }
+
+  @override
+  Future<void> setThemeMode(IntercomTheme theme) async {
+    await _channel.invokeMethod('setThemeMode', {'theme': theme.name});
   }
 
   /// Convert the [PlatformException] details to [IntercomError].