+++ intercom_flutter_web/lib/intercom_flutter_web.dart
@@ -3,14 +3,15 @@ import 'dart:js_interop';
 import 'dart:js_interop_unsafe';
 
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';
+import 'package:intercom_flutter_platform_interface/enumeral.dart';
 import 'package:intercom_flutter_platform_interface/intercom_flutter_platform_interface.dart';
 import 'package:intercom_flutter_platform_interface/intercom_status_callback.dart';
 import 'package:uuid/uuid.dart';
 import 'package:web/web.dart' as web;
 
 /// export the enum [IntercomVisibility]
-export 'package:intercom_flutter_platform_interface/intercom_flutter_platform_interface.dart'
-    show IntercomVisibility;
+export 'package:intercom_flutter_platform_interface/enumeral.dart'
+    show IntercomVisibility, IntercomTheme;
 export 'package:intercom_flutter_platform_interface/intercom_status_callback.dart'
     show IntercomStatusCallback, IntercomError;
 
@@ -216,7 +217,13 @@ class IntercomFlutterWeb extends IntercomFlutterPlatform {
     // shutdown will effectively clear out any user data that you have been passing through the JS API.
     // but not from intercomSettings
     // so manually clear some intercom settings
-    removeIntercomSettings(['user_hash', 'user_id', 'email']);
+    removeIntercomSettings([
+      'user_hash',
+      'intercom_user_jwt',
+      'user_id',
+      'email',
+      'auth_tokens',
+    ]);
     // shutdown
     globalContext.callMethod('Intercom'.toJS, 'shutdown'.toJS);
     print("logout");
@@ -324,6 +331,39 @@ class IntercomFlutterWeb extends IntercomFlutterPlatform {
     }
 
     return {};
+  }
+
+  @override
+  Future<void> setUserJwt(String jwt) async {
+    globalContext.callMethod(
+      'Intercom'.toJS,
+      'update'.toJS,
+      updateIntercomSettings('intercom_user_jwt', jwt).jsify(),
+    );
+    print("jwt added");
+  }
+
+  @override
+  Future<void> setAuthTokens(Map<String, String> tokens) async {
+    globalContext.callMethod(
+      'Intercom'.toJS,
+      'update'.toJS,
+      updateIntercomSettings('auth_tokens', tokens).jsify(),
+    );
+    print("Auth tokens added");
+  }
+
+  @override
+  Future<void> setThemeMode(IntercomTheme theme) async {
+    globalContext.callMethod(
+      'Intercom'.toJS,
+      'update'.toJS,
+      updateIntercomSettings(
+        'theme_mode',
+        theme == IntercomTheme.none ? null : theme.name,
+      ).jsify(),
+    );
+    print("Theme overridden");
   }
 
   /// get the [window.intercomSettings]