+++ intercom_flutter_platform_interface/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## 2.0.6
+
+* Added method `setThemeMode`.
+
+## 2.0.5
+
+* Deprecated `handlePushMessage`
+
+## 2.0.4
+
+* Added method `setAuthTokens`.
+
+## 2.0.3
+
+* Added method `setUserJwt`.
+
 ## 2.0.2
 
 * Added method `isUserLoggedIn`.